Resolve "Loading a default save file"
Added: A boolean method to check if previous game is present. A way to run previous game if present. A method that returns summary string about previous game.  A temporary previousGame.xml file to test these features, A saves directory, for future save files and for the previousGame.xml file.